moved the utils module to a separate directory
