+++ src/mock/event.js
@@ -1,5 +1,5 @@
 import dayjs from "dayjs";
-import {getRandomInteger} from "../utils";
+import {getRandomInteger} from "../utils/utils";
 
 const getDescription = () => {
   const description = [
