+++ src/utils/utils.js
@@ -1,4 +1,4 @@
-import {HOUR, DAY} from "./const";
+import {HOUR, DAY} from "../const";
 
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
