+++ src/view/event-list.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/utils";
 
 const createEventList = () => {
   return `<ul class="trip-events__list"></ul>`;
