+++ src/view/event-offer.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/utils";
 
 const renderOffer = (event) => {
   return event.offers
