+++ src/view/no-event-item.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/utils";
 
 const noEventItem = () => {
   return `<p class="trip-events__msg">Click New Event to create your first point</p>`;
