+++ src/view/trip-cost.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/utils";
 
 const createTripCost = (events) => {
   const totalPrice = events.reduce((price, event) => {
