+++ src/view/trip-tabs.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/utils";
 
 const createTripTabs = () => {
   return `<nav class="trip-controls__trip-tabs  trip-tabs">
